requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg[binary]==3.3.2
python-dotenv==1.0.1
pydantic==2.6.4
openai==1.30.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

streamlit==1.32.0
